Indoor-Localization-main/requirements.txt
numpy>=1.19.2
tensorflow-cpu>=2.4.0
matplotlib>=3.3.2
seaborn>=0.11.0
scikit-learn>=0.24.0
pandas>=1.1.3
jupyter>=1.0.0
pytest>=6.0.0
black>=21.0
flake8>=3.9.0
joblib>=1.0.0
tqdm>=4.50.0 